backend/requirements.txt
# Core
fastapi==0.104.1
uvicorn[standard]==0.29.0
python-dotenv==1.0.1
pydantic==2.7.1

# Database
psycopg2-binary==2.9.9

# AI / Agents
crewai==0.11.2
langchain-google-genai==1.0.3
google-generativeai==0.5.4

# Scraping (optional)
praw==7.7.1
pytrends==4.9.2
google-api-python-client==2.127.0

# Integrations
sendgrid==6.11.0
requests==2.31.0

# Scheduling
apscheduler==3.10.4

# Frontend
streamlit==1.33.0

# Dev
httpx==0.27.0